% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{GET STATS AND PLOT}
\usage{
get_stats(
  mot1_MTTF = 8000,
  mot2_MTTF = 12000,
  prop1_MTTF = 300,
  prop2_MTTF = 500,
  bat1_t = 800,
  bat1_lc = 400,
  bat2_t = 1090,
  bat2_lc = 1000,
  lambda_gc = 2 * 10^-6,
  lambda_main = 2.77 * 10^-6,
  lambda_nav = 9.41 * 10^-6,
  lambda_es = 5.01 * 10^-6,
  lambda_pay = 1.1 * 10^-6
)
}
\arguments{
\item{mot1_MTTF}{....}

\item{mot2_MTTF}{....}

\item{prop1_MTTF}{....}
}
\description{
This function will calculate the reliability of a commercial drone system.
It splits a commercial drone into six different systems: Ground Control,
Mainframe, Power System, Payload, Electronics System, and Navigation Systems.
We further broke down the Power System into 3 main components, the battery,
motors, and propellers. The failure rates for the 6 systems were derived from
literary research and the mean time to failure of the batteries, motors, and
propellers were found by using the most common commercial options.

The MTTF for the original motor and propellers (mot1 and mot2) are the lower ends
of the ranges that we found online for drone motors and propellers. The MTTF for
the improved motors and propellers, (mot2 and prop2) are the higher ends of the 
ranges that we found online. Using these values, we are comparing the reliability
as a result of using higher quality components. In this function you are free to
enter your own MTTF for intiial and final motor/battery components and observe
how this affects a drone's reliability.

For the batteries, our initial battery component (bat1) is a lithium ion battery
system. We found its MTTF by multiplying the average lifetime of one cycle (bat1_t)
by the
expected life cycles.
}
\author{
Hengtao Zhao & Andres Garcia
}
